new namespace for DBPFEntries
